+++ src/NetAppFiles/NetAppFiles/Pool/SetNetAppFilesPool.cs
@@ -108,7 +108,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Pool
         [Parameter(
             Mandatory = false,
             HelpMessage = "Maximum throughput in MiB/s that can be achieved by this pool and this will be accepted as input only for manual qosType pool with Flexible service level")]
-        public double? CustomThroughputMibps { get; set; }
+        public double? CustomThroughput { get; set; }
 
         [Parameter(
             ParameterSetName = ParentObjectParameterSet,
@@ -176,7 +176,7 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Pool
                 Tags = tagPairs,
                 QosType = QosType,
                 CoolAccess = CoolAccess,
-                CustomThroughputMibps = CustomThroughputMibps
+                CustomThroughputMibps = CustomThroughput
             };
 
             if (ShouldProcess(Name, string.Format(PowerShell.Cmdlets.NetAppFiles.Properties.Resources.UpdateResourceMessage, ResourceGroupName)))