+++ src/NetAppFiles/NetAppFiles/ChangeLog.md
@@ -20,7 +20,7 @@
 ## Upcoming Release
 * Updated to api-version 2025-06-01
 * Added `Flexible` to `ServiceLevel` for `New-AzNetAppFilesPool`
-* Added `CustomThroughputMibps` to `New-AzNetAppFilesPool` and `Update-AzNetAppFilesPool`
+* Added `CustomThroughput` to `New-AzNetAppFilesPool` and `Update-AzNetAppFilesPool`
 * Added `AcceptGrowCapacityPoolForShortTermCloneSplit` to `New-AzNetAppFilesVolume`
 * Added `Split-AZNetAppFilesCloneFromParent` cmdlet to split a clone from its parent volume